requirements.txt
Django==4.1.4
djangorestframework==3.14
djangorestframework-simplejwt==5.2.2